Fix mdlint error in support-matrix doc (#3684)
The support matrix documentation was recently updates to remove
references to cgroupv1 limitations. For some reason the mdlint check did
not run on that change, and there was a minor issue with its formatting.

This removes an extra blank line from the end of the file to make the
linter happy.

Signed-off-by: Sean McGinnis <smcginnis@vmware.com>
(cherry picked from commit 44f2bc40b7a4c67603d2bda4bf333060235225d6)