urllint: use unbuffered async channel (#3807)
We are seeing periodic failures checking URLs, even though these URLs
are valid, well-known sites such as sonobuoy.io. It's possible that even
with the small number of concurrent requests we make, that could be
triggering some sort of DoS protection from these sites.

This changes the async queue we use for processing results to be
unbuffered. It had been a buffered queue of 2, so we could have multiple
GET checks happening simultaneously. An unbuffered queue will block the
from moving on to the next check until the result is picked up to be
processed. Even with this unbuffered change, the speed is just over 4
seconds to check all URLs versus the original 60 second time seen when
all checks were fully sequentual.

Signed-off-by: Sean McGinnis <smcginnis@vmware.com>