Don't skip linting on test code
We had been skipping running linters on code in test directories. There
really isn't any reason we shouldn't make sure our test code is as good
of quality as our functional code, so this removes that skip to make
sure this is covered. Thankfully it did not uncover hidden problems, but
it will help make sure nothing is introduced going forward.

Also removes the golangci-lint version from the config file. This is now
controlled by our test tool go.mod entry, so no need to have that in the
config file.

Signed-off-by: Sean McGinnis <smcginnis@vmware.com>