Optimize url check to just use HEAD
Our urllinter currently performs a GET request on each URL found in the
repo. Since that will actually return the full page content of the site,
and all we need to know is if the page exists, this results in data
being sent over the wire that we just end up ignoring.

This changes the link checking to use a HEAD request instead. That will
tell us if the page exists or not, while not sending any unnecessary
data with the response.

For whatever reason, the k8s testgrid site does not allow HEAD requests
and returns a 405. This is the only site out of the list of URLs that
behaves this way. To allow us to use the more optimal HEAD request for
the majority of sites, this will detect if a 405 is returned and retry
the link check using the full GET request. This makes the checking
slightly more robust.

Also sets the user-agent on these requests so there is some indication
where all these requests are coming from in case a site operator notices
a burst of requests and wants to investigate.

Signed-off-by: Sean McGinnis <smcginnis@vmware.com>