Clean up configured linters in golangci-lint
We have a long list of linters configured in our golangci-lint
configuration. Since the linting of all modules takes a noticeable
amount of time, I went through the list to see if there were any that
were redundant or of very low value to try to optimize this a little.

This removes a few of the configured linters:

- depguard - requires configuration that we are not using
- dogsled - checks for too many blank assignments (_, _ = x())
- gosimple - used to simplify code when using "fix", which we don't
- rowserrcheck - used for checking SQL query error checks
- unused - already covered by deadcode

Signed-off-by: Sean McGinnis <smcginnis@vmware.com>