UI: Fix docker build of UI files
With the addition of generating client code from the swagger spec, the
directory path we were mounting in the docker file to perform the site
build wasn't able to reach the swagger spec.

This changes where we mount the local directory and run the build
scripts so it includes the path to the spec.

Also fixed the logic in the root plugin build to properly recognize if
the site has been built.

Signed-off-by: Sean McGinnis <smcginnis@vmware.com>