Resolve minikube image via kubernetes version (#3981)
This commit moves (back) to minikube resolution based on the kubernetes
version specified in the TKr. This is needed because there are unique
images for each driver type (e.g. kvm2 and docker).

Until we setup our TKrs to express many minikube images, we will stick
to this model and leverage default images resolved by minikube.

Signed-off-by: joshrosso <rossoj@vmware.com>