cartographer: consume release contents from tgz
given that the repository where does bits come from
(vmware-tanzu/package-for-cartographer) is aimed towards the generation
of packages for multiple projects targetted at multiple distributions
(e.g., TAP _and_ TCE), now for each projects and distribution, a `.tgz`
is included in the release so that `package.yaml` for each can be
fetched.

note that previous releases (0.2.2) have been updated to conform to the
same pattern.

Signed-off-by: Ciro S. Costa <ciroscosta@vmware.com>