Fix test badge link in README
We have a badge in our README file that is supposed to show the current
status of tests. Since that was added, we renamed and split out several
tests, so the status it was referring to no longer exists. This updates
that badge to point to our current job that runs all tests.

Signed-off-by: Sean McGinnis <smcginnis@vmware.com>