UI: more API implementation
This is still laying the foundation for our API. It adds some more
foundation, with implementation for some of the common system info calls
and the Docker provider.

The architecture is still evolving as the implementation is being worked
through, rather than trying to set something up front. So there are
likely more changes.

Some of the basic GET endpoints have been tested locally, but the
deployment calls have not been fully tested yet.

Signed-off-by: Sean McGinnis <smcginnis@vmware.com>