Blog: make doc links go to specific version
The versioned docs changed where some of the referenced links point to.
In many cases the redirect will get to the correct page, but there were
a few links in our blogs that were now resulting in 404s.

This updates the links in a couple blog posts to get to specific
versions. This fixes and link and makes sure future modifications will
not break existing posts.

Signed-off-by: Sean McGinnis <smcginnis@vmware.com>