Enable `make test-plugins`
We had a placeholder target in the root Makefile for a `test-plugins`
target. This was meant to run unit test coverage for our CLI plugins.

We have since added unit test coverage for our plugins, so now is a good
time to finish wiring this up. This updates our plugins to support their
expected `make test` target, and `make test-plugins` will now invoke
this for all CLI plugins in the repo.

Also added this to the `check-all` GitHub Action so unit tests of the
CLI plugins are included as part of our comprehensive testing.

Signed-off-by: Sean McGinnis <smcginnis@vmware.com>