Enable `make test-packages`
We've encouraged package authors to have test coverage, but so far we
have not had a common way to run any of these tests. This wires up the
`test-packages` target to run unit testing for all packages in the repo.
It also adds a GitHub Action so these tests are run whenever there are
modifications to the `addons/packages` files.

Signed-off-by: Sean McGinnis <smcginnis@vmware.com>