Add timeout to cache restore in Actions (#4225)
We have been seeing the restore of cache files basically hang during the
restore process, causing it to eventually timeout after 6 hours of
inactivity. This fails the entire job which then needs to be rerun.

This change sets a 10 minute timeout for the cache restore. Under normal
conditions, the restore is fairly quick. So if it's taken 5 minutes, we
know something isn't quite right.

It also sets the flag to tell the action to continue on this step
failure. It's better to have decreased performance from cache misses
than to fail if it's not present.

Signed-off-by: Sean McGinnis <smcginnis@vmware.com>