Fix "dependabot[bot]" name for auto PR merge (#4308)
We added an action to automatically merge dependabot updates only for
the webhook, since that is fairly safe and happens often. There was a
small issue with the script in that the "Dependabot" account name is
actually "dependabot[bot]", causing the name not to match and the
merging to be skipped. This updates the name to match.

Signed-off-by: Sean McGinnis <smcginnis@vmware.com>