Update doc links to "edge" instead of "main"
With recent changes in the versioned docs to allow seeing locally made
changes, the "main" version of the docs are no longer published and
there is just "edge". We had a few links in our README and CONTRIBUTING
docs that referenced other documentation under this path that now fail.

This updates those links to point to their "edge" equivalents.

Signed-off-by: Sean McGinnis <smcginnis@vmware.com>