Fix unmanaged cluster log verbosity enforcement
There were issues with the way log verbosity was being set which caused
the incorrect level to be used, resulting in some expected lines not
printing at the default log level.

This updates the handling to correctly parse and set the level from the
flag. Also some minor refactoring to make it a little easier to
understand.

Signed-off-by: Sean McGinnis <smcginnis@vmware.com>