UI: Explicitly set edition for cluster creation
This sets the Edition parameter for cluster creation to be hard coded to
"TCE" for now. If this extension is moved to work with multiple editions
we will need to update this to check config settings, but for now we
know it is only going to be used with the community edition version so
we can just set it directly.

Signed-off-by: Sean McGinnis <smcginnis@vmware.com>