UI: fix launching browser to UI path
The current browser launch was opening the root server path. The
redirect to the /ui path was not working, so this meant by default a
user would get sent to a 404 page and then need to add the /ui on the
end to actually see the UI.

This fixes both the initial launch URL to go directly to the UI as well
as the redirect so if someone navigated to the root, they would actually
get redirected to the right location.

Signed-off-by: Sean McGinnis <smcginnis@vmware.com>