Fix doc issue link and remove edit option
The docs pages have a "Report Issue" link in the right nav to quickly
get someone over to the docs template for filing any issues they see in
the documentation. This had been attempting to put a link in the title
based on the title of the page, but for most of our pages this ends up
being an empty value.

This updates the query to just put the relative link to the page
directly in the issue title, and puts a placeholder to try to nudge the
user to add an actual description to it.

While looking at this it was also noticed that we had an "Edit" link
there that would attempt to use the GitHub UI editor to be able to make
changes and submit a new PR. We had disabled this ability, so the link
would actualy bring the user to a GitHub 404 page.

Since the ability to file an issue is sufficient, to address the Edit
404 the link is just removed.

Signed-off-by: Sean McGinnis <smcginnis@vmware.com>