Allows kube-system to be orphaned when deleting metrics-server
When the metrics-server is installed it might add the kube-system namespace
to its list of resources. Subsequently, when the metrics-server is deleted,
it will try to delete the kube-system namespace, but this is not allowed so
package deletion fails.

This patch makes it possible to delete the metrics-server package without
having to delete the kube-system namespace.