UI: Restore build output README to avoid lint failures
This restores the README file from our web build output folder. This is
needed so the folder is present when the linter runs in CI. The lint job
does not do a full build first, so the generated files are not present.
Without something there the embed directive in the go source will fail
because there will be no files found.

Signed-off-by: Sean McGinnis <smcginnis@vmware.com>