UI: Initialize clients on initial launch (#4811)
There are two clients used for performing API operations, a TKGClient
and a TkgClient. We were initializing them at the point they were
needed.

Since almost all operations need to use one of the clients, and to allow
us to fail immediately if there is some sort of problem preventing the
creation of these clients, this initializes the clients right away as we
start, before we launch the UI, so we can make sure it is successful and
error out with a reasonable error message so the user doesn't go through
any wizards only to find out nothing works.

We also had an issue where the `~/.config/tanzu` directory was not being
initialized. We need that to exist along with the BOM and configuration
information that gets populated under it. Turns out this is triggered by
the creation of the TKGClient. It checks if the directory exists and
will perform the necessary steps to intialize everything.

Errors were seen when trying to create a management cluster on AWS. This
may be due to the missing content under the config directory.

Signed-off-by: Sean McGinnis <smcginnis@vmware.com>