UI: Centralize init of default CAPI providers
In addition to the infrastructure provider, there are a set of other
providers that need to be specified when creating a cluster. This was
proven out by making changes specifically for the AWS provider, but the
same thing needs to happen for all other infrastructure.

To support this, this moves the discovery of those defaults from the BOM
file into the app-wide initialization so all infra provider workflows
have easy access to these values.

Signed-off-by: Sean McGinnis <smcginnis@vmware.com>