UI: Centralize init of default CAPI providers (#4879)
* UI: Centralize init of default CAPI providers

In addition to the infrastructure provider, there are a set of other
providers that need to be specified when creating a cluster. This was
proven out by making changes specifically for the AWS provider, but the
same thing needs to happen for all other infrastructure.

To support this, this moves the discovery of those defaults from the BOM
file into the app-wide initialization so all infra provider workflows
have easy access to these values.

* UI: Change default timeout for long running ops

The default timeout on our client was set very low (30 seconds). This is
not long enough for operations like management cluster creation. This
changes the default to be our long running operation default of 30
minutes.

We may want to make this dynamic in the future.

Signed-off-by: Sean McGinnis <smcginnis@vmware.com>