Skip sudo on install of running as root (#4960)
Our install script attempts to figure out the best place to install the
`tanzu` binary to. It will look in the user's home directory for a
`~/bin` path, and if that is not found it will try to install to the
common path of `/usr/local/bin`.

In the latter case, in normal cases this needs to be run using sudo to
have permissions to install to the system-wide path. But when running as
root, the use of sudo is a noop since there are no elevated permissions
needed.

On most systems, when running as root the `sudo` command is still
present somewhere in `$PATH`. But some systems do not have this
available. So despite the fact that the root user doesn't actually need
`sudo`, the command fails due to the command not being found.

To avoid this situation, this changes the install script to detect if we
are running as root, and if so, run the install to the system-wide path
without trying to invoke using sudo.

Signed-off-by: Sean McGinnis <smcginnis@vmware.com>